Add tests for oembed field
