+++ ui/js/dfv/pods-dfv.min.asset.json
@@ -1,1 +1,1 @@
-{"dependencies":["wp-api-fetch","wp-autop","wp-components","wp-compose","wp-data","wp-element","wp-i18n","wp-keycodes","wp-url"],"version":"7761fca647c373b151ccaeeaf8929ca9"}
+{"dependencies":["wp-api-fetch","wp-autop","wp-components","wp-compose","wp-data","wp-element","wp-i18n","wp-keycodes","wp-url"],"version":"71c3abf4577b70bfddb39208597a896a"}
